allocator: allow task-id vectors to handle multiple task instances
Improve the handling of *deployments* in allocator. A mesomatic
task-info already extends the notion of mesos task-info by
allowing the ~count~ and ~maxcol~ fields to be present, which
respectively indicate the number of instances of a task to
start and the maximum number of instances to colocate.

By allowing this, a single TaskInfo provided to the allocator will
result in multiple TaskInfo ready to be launched, making it hard
for frameworks to keep track of IDs.

This commit allows a vector associated with the ~task-id~ field
of TaskInfo, expected to match ~count~'s size to be provided,
each generated TaskInfo in the allocator's result will have
individual task ids picked from the vector.