`cond` -> `case` where appropriate (#21)
* `cond` -> `case` where appropriate

- Moving from `cond` to `case` when possible turns a worst-case O(N)
(where N is the number of possible objects to which the test constant
might be equal) check into a worst-case O(1) check.
- While it is not clear whether the affected code paths are
bottlenecks, performance will nevertheless be improved in those code
paths.
- All tests continue to pass

* `case` -> dispatch map for enums

Following a comment by @mpenet , `case` dispatch on enums was changed
to a dispatch map whose keys are enums, which is accessed via a call to
`get`. Not O(1) like `case`, but still O(log32(N)), which is quite
comparable.

* dispatch-map -> `case-enum`

As per @mpenet , added/pasted in the `case-enum` function and avoided
the dispatch maps. Looks cleaner.